fix pkg_mv_plugins to only try and move things if  is not a symlink

+++ eclass/nsplugins.eclass
@@ -1,6 +1,6 @@
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/eclass/nsplugins.eclass,v 1.10 2003/02/28 09:15:04 vapier Exp $
+# $Header: /var/cvsroot/gentoo-x86/eclass/nsplugins.eclass,v 1.11 2003/03/22 05:58:24 azarah Exp $
 #
 # Author: Martin Schlemmer <azarah@gentoo.org>
 #
@@ -32,7 +32,7 @@ src_mv_plugins() {
 pkg_mv_plugins() {
 
 	# Move old plugins dir
-	if [ -d ${ROOT}/$1 ]
+	if [ -d "${ROOT}/$1" -a ! -L "${ROOT}/$1" ]
 	then
 		mkdir -p ${ROOT}/usr/lib/${PLUGINS_DIR}
 		cp -a ${ROOT}/$1/* ${ROOT}/usr/lib/${PLUGINS_DIR}
@@ -47,7 +47,7 @@ inst_plugin() {
 	MYFILE="`echo ${1} | gawk  -F '/' '{ print $NF }'`"
 
 	# Install the plugin if none is installed
-	if [ ! -L /usr/lib/${PLUGINS_DIR}/${MYFILE} ]
+	if [ ! -L "/usr/lib/${PLUGINS_DIR}/${MYFILE}" ]
 	then
 		dodir /usr/lib/${PLUGINS_DIR}
 		# $ROOT should only be used in pkg_*() functions ...
